% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbetamupsi_abstract.r
\name{findbetamupsi_abstract}
\alias{findbetamupsi_abstract}
\title{The findbetamupsi (abstract) function}
\usage{
findbetamupsi_abstract(themean.cat, thevariance.cat,  psi.percentile=0.90, percentile.median, percentile95value)
}
\arguments{
\item{themean.cat:}{specify your prior belief about the mean. It takes a value among c("Very low","Low","Average","High","Very high").}

\item{thevariance.cat:}{specify your prior belief about the variance. It takes a value among c("Very low","Low","Average","High","Very high").}

\item{psi.percentile:}{specify the level of confidence that a certain fraction of the units under study has a prevalence less than the percentile.median. It takes a value between 0 and 1 and the default is 0.90.}

\item{percentile.median:}{specify the median value that corresponds to the defined psi.percentile. It takes a value between 0 and 1 and has to be higher than both themean and the percentile.}

\item{percentile95value:}{specify the value that the percentile.median does not exceed with 95\% confidence. It takes a value between 0 and 1 and has to be higher than the percentile.median.}

\item{parameters:}{The beta distribution parameters Beta(a,b)}

\item{bot_param:}{simulated mu and psi of Beta(mu psi,psi(1-mu))}

\item{summary:}{A basic summary of the elicited prior}

\item{input:}{The initial input value that produced the above prior.}
}
\description{
A function to estimate (a) the parameters of a Beta distribution for the expected mean of a proportion - usually the prevalence of disease/infection for the units in an area/region and (b) the parameters of a Gamma distribution expressing our prior beleif about the variability of the prevalence estimates across the units of the area/region under consideration.
}
\examples{
## Example
## The mean prevalence of a disease/infection for the units within an area/region
## is thought to be generally low and its variance is neither high nor low,
## we are also confident that 90\% of all units have a prevalence
## less or equal to 0.60 and we are 95\% certain that it does not exceed 0.70

findbetamupsi_abstract(themean.cat="Low", thevariance.cat="Average", psi.percentile=0.90, percentile.median=0.60, percentile95value=0.70)

}
\references{
Branscum, A. J., Gardner, I. A., & Johnson, W. O. (2005): Estimation of diagnostic test sensitivity and specificity through Bayesian modeling. Preventive veterinary medicine, \bold{68}, 145--163.
}
