% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbeta_plot.R
\name{plot.PriorGen}
\alias{plot.PriorGen}
\title{The findbeta plot function}
\usage{
function(findbeta.object, lines=FALSE, ...)
}
\arguments{
\item{findbeta.object:}{An object of type findbeta produces of one of the other PriorGen functions.}

\item{lines:}{If lines=TRUE then it plots over the last plotted plot.}

\item{...:}{Additional base plot parameters.}

\item{parameters:}{The beta distribution parameters Beta(a,b)}

\item{summary:}{A basic summary of the elicited prior}

\item{input:}{The initial input value that produced the above prior.}
}
\description{
A function that plots any object of the class findbeta.
}
\examples{
##Example 1
##Based on the available literature the mean value for the sensitivity of a test
##is expected to be generally low and its variance not that low but not that much neither.

res_abs_1<-findbeta_abstract(themean.cat="Low", thevariance.cat="Average")
findbeta_plot(res_abs_1,main="Plot of the findbeta_abstract function",lwd=3)

}
\references{
Branscum, A. J., Gardner, I. A., & Johnson, W. O. (2005): Estimation of diagnostic test sensitivity and specificity through Bayesian modeling. Preventive veterinary medicine, \bold{68}, 145--163.
}
