% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorGen-package.R
\docType{package}
\name{PriorGen}
\alias{PriorGen}
\title{Generates Prior Distributions for Proportions}
\description{
Translates beliefs into prior information in the form
of Beta and Gamma distributions. It can be mainly used
for the generation of priors on the prevalence of disease
and the sensitivity/specificity of diagnostic tests.
}
